+++ src/main/java/thelibrary/api/biblioteca/entity/Livro.java
@@ -1,6 +1,5 @@
 package thelibrary.api.biblioteca.entity;
 
-import jakarta.persistence.Embedded;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
@@ -8,7 +7,6 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.ISBN;
-import thelibrary.api.biblioteca.entity.enumerable.GeneroLiterario;
 
 import javax.persistence.*;
 
@@ -21,7 +19,7 @@ import javax.persistence.*;
 public class Livro {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @NotNull
     private String titulo;
     private Autor autor;
